поправил по замечаниям
