+++ src/components/form-review/form-review.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, FC, useState } from 'react';
 
 export const FormReview: FC = () => {
   const [formData, setFormData] = useState({
-    rating: '',
+    rating: -1,
     review: ''
   });
 
@@ -11,93 +11,29 @@ export const FormReview: FC = () => {
     setFormData({...formData, [name]: value});
   };
 
+  const raitings = [
+    { value: 5, label: 'perfect'},
+    { value: 4, label: 'good'},
+    { value: 3, label: 'not good'},
+    { value: 2, label: 'badly'},
+    { value: 1, label: 'terribly'}
+  ];
   return(
     <form className='reviews__form form' action='#' method='post'>
       <label className='reviews__label form__label' htmlFor='review'>
         Your review
       </label>
       <div className='reviews__rating-form form__rating'>
-        <input
-          className='form__rating-input visually-hidden'
-          name='rating'
-          //defaultValue={5}
-          id='5-stars'
-          type='radio'
-        />
-        <label
-          htmlFor='5-stars'
-          className='reviews__rating-label form__rating-label'
-          title='perfect'
-        >
-          <svg className='form__star-image' width={37} height={33}>
-            <use xlinkHref='#icon-star' />
-          </svg>
-        </label>
-        <input
-          className='form__rating-input visually-hidden'
-          name='rating'
-          defaultValue={4}
-          id='4-stars'
-          type='radio'
-        />
-        <label
-          htmlFor='4-stars'
-          className='reviews__rating-label form__rating-label'
-          title='good'
-        >
-          <svg className='form__star-image' width={37} height={33}>
-            <use xlinkHref='#icon-star' />
-          </svg>
-        </label>
-        <input
-          className='form__rating-input visually-hidden'
-          name='rating'
-          defaultValue={3}
-          id='3-stars'
-          type='radio'
-        />
-        <label
-          htmlFor='3-stars'
-          className='reviews__rating-label form__rating-label'
-          title='not bad'
-        >
-          <svg className='form__star-image' width={37} height={33}>
-            <use xlinkHref='#icon-star' />
-          </svg>
-        </label>
-        <input
-          className='form__rating-input visually-hidden'
-          name='rating'
-          defaultValue={2}
-          id='2-stars'
-          type='radio'
-        />
-        <label
-          htmlFor='2-stars'
-          className='reviews__rating-label form__rating-label'
-          title='badly'
-        >
-          <svg className='form__star-image' width={37} height={33}>
-            <use xlinkHref='#icon-star' />
-          </svg>
-        </label>
-        <input
-          className='form__rating-input visually-hidden'
-          name='rating'
-          id='1-star'
-          type='radio'
-          onChange={handleFieldChange}
-          value={formData.rating}
-        />
-        <label
-          htmlFor='1-star'
-          className='reviews__rating-label form__rating-label'
-          title='terribly'
-        >
-          <svg className='form__star-image' width={37} height={33}>
-            <use xlinkHref='#icon-star' />
-          </svg>
-        </label>
+        {raitings.map(({ value, label }) => (
+          <>
+            <input className="form__rating-input visually-hidden" name="rating" id={`${value}-stars`} type="radio" value={value} onChange={handleFieldChange}/>
+            <label className="reviews__rating-label form__rating-label" htmlFor={`${value}-stars`} title={label}>
+              <svg className ="form__star-image" width="37" height="33">
+                <use xlinkHref="#icon-star"></use>
+              </svg>
+            </label>
+          </>
+        ))}
       </div>
       <textarea
         className='reviews__textarea form__textarea'
@@ -105,7 +41,6 @@ export const FormReview: FC = () => {
         name='review'
         placeholder='Tell how was your stay, what you like and what can be improved'
         onChange={handleFieldChange}
-        value={formData.review}
       />
       <div className='reviews__button-wrapper'>
         <p className='reviews__help'>
@@ -117,7 +52,7 @@ export const FormReview: FC = () => {
         <button
           className='reviews__submit form__submit button'
           type='submit'
-        //  disabled
+          disabled={formData.rating < 1 || formData.review.length < 50 }
         >
           Submit
         </button>
